Brachinous defeat
The brachinous enemy can now interact with hazards to get stunned. The time it is stunned is infinite but can be temporal just removing a comment inside it's collision detection script.
Enemies are now inside 2 groups: Hazards and Enemies. This was done so the Collision event can distinguish enemies and hazards. While the enemy is stunned, it cannot harm the player. This last bit hasn't been applied to the 2 previous enemies but it can be implemented just adding and removing them from the "Hazard" group.