Enemy Patrolling Updated
The Circle Movement Area has been replaced by a Path2D which is is an array of position that the enemy can follow randomally or secuentially. Every enemy has a Curve2D pre established but it can be modified manually. Changing the type of movement is done inside the main enemy script via an exported variable called "Movement Type".